[ Version 2.6.8.1 ] * *Bug* - Fixed notices errors. * *Bug* - Fixed extra tr error. * *Bug* - Fixed issue where text was not being put in the right place on a mDocs list, now content will display as published on all mDocs lists. * *Bug* - Fixed error if file was not found.
git-svn-id: https://plugins.svn.wordpress.org/memphis-documents-library/trunk@1010744 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ mdocs-settings-page.php
@@ -65,7 +65,7 @@ function mdocs_settings($cat) {
 		<td>
 			<?php
 			$wp_roles = get_editable_roles(); 
-			$mdocs_roles = get_option('mdocs-view-private');
+			$mdocs_roles = mdocs_init_view_private();
 			foreach($wp_roles as $index => $role) {
 			?>
 			<input type="checkbox" name="mdocs-view-private[<?php echo $index; ?>]" value="1"  <?php checked($mdocs_roles[$index] , 1) ?> /> <span><?php echo $role['name']; ?></span><br>
@@ -316,7 +316,7 @@ function mdocs_filesystem_cleanup_submit() {
 	$cleanup = mdocs_filesystem_cleanup_init();
 	$upload_dir = wp_upload_dir();
 	foreach($cleanup['files'] as $file) {
-		unlink($file);
+		if(is_file($file)) unlink($file);
 	}
 	foreach($cleanup['data'] as $data) {
 		if(isset($data['id'])) wp_delete_attachment( intval($data['id']), true );
